# Hash / HashWithIndifferentAccess speed improvements