@rafaelfranca 2019-03-27
Thank you for the pull request. For future reference, can you show the numbers on how fast this is comparing with the previous implementation?