@kaspth 2019-03-27
Some memory profiling would be interesting to see too