@kaspth 2019-03-27
https://github.com/SamSaffron/memory_profiler is one way to generate those reports. Especially this new version seems like it generates lots of intermediate objects.