@timoschilling 2019-03-27
@kaspth Do you think that because of `convert_key`? I assume that didn't make a difference because of the old implementation uses `delete` which uses `convert_key` too.