@kaspth 2019-03-27
I'm thinking about memory because `slice`, `self.keys`, `-`, `keys` and `map` all create intermediate arrays.