@timoschilling 2019-03-27
I added the profile results based on the test methods from the benchmark.