@rafaelfranca 2019-03-28
HashWithIndifferentAccess tests seems to be broken.