@timoschilling 2019-04-02
I fixed the specs and add a optimisation for `HashWithIndifferentAccess#values_at` too.